requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
gast==0.4.0
glcontext==2.3.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.0
six==1.16.0
tensorboard==2.8.0
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3

